feat: Add Browse component and related changes
This commit adds the new `Browse.jsx` component and related changes to the project. The `Browse` component renders a simple "Browse" text and is imported in the `App.jsx` file. A new file `index.js` is also created in the `Components` folder to export the `Browse` component. Additionally, the new `Body.jsx` component is added with its import and rendering in the `App.jsx` file. The `Body` component uses the `react-router-dom` library for routing and defines routes for the `Login` and `Browse` components. The routing is set up with the `createBrowserRouter` and `RouterProvider` components inside `Body.jsx`.

The project is also updated with new dependencies. The `firebase` library is added with version "10.13.1". The `react-router-dom` dependency is upgraded to version "6.26.1".

New utility files `validate.jsx` and `firebase.jsx` are also added to the project. `validate.jsx` contains a function `checkValidData` used for form data validation, and `firebase.jsx` contains the configuration for Firebase that is imported in other files.

Lastly, the "README.md" file is updated to reflect the changes made in the project.